requirements/base.txt
# Core dependencies for the DailyAssist application

# AWS and Chalice
boto3>=1.26.0
chalice>=1.27.0

# Database
pynamodb>=5.2.0

# Date handling
python-dateutil>=2.8.2

# Data validation
pydantic>=2.0.0

# Typing support
typing-extensions>=4.5.0
